+++ lib/app/routes/app_pages.dart
@@ -1,5 +1,4 @@
 import 'package:get/get.dart';
-import '../modules/calculator/bindings/calculator_binding.dart';
 import '../modules/calculator/views/calculator_getx_view.dart';
 import '../modules/history/bindings/history_binding.dart';
 import '../modules/history/views/history_getx_view.dart';
@@ -12,11 +11,7 @@ class AppPages {
   static const initial = Routes.calculator;
 
   static final routes = [
-    GetPage(
-      name: _Paths.calculator,
-      page: () => const CalculatorView(),
-      binding: CalculatorBinding(),
-    ),
+    GetPage(name: _Paths.calculator, page: () => const CalculatorView()),
     GetPage(
       name: _Paths.history,
       page: () => const HistoryView(),